Refactor ngSCTP: migrate multi-homing to SctpElement, remove duplicate NgSctpService
- Migrated `PathState`, `PathStatus` logic to `SctpElement`.
- Added `paths` and `congestionControl` to `SctpElement`.
- Implemented `failover`, `recoverPath`, and `primaryPath` handling in `SctpElement`.
- Deleted obsolete `libs/couch/src/commonMain/kotlin/borg/trikeshed/couch/ccek/NgSctpService.kt`
- Deleted obsolete `src/commonMain/kotlin/borg/trikeshed/ccek/NgSctpService.kt`
- Executed `libs/ngsctp` tests successfully.
