+++ libs/ngsctp/src/commonMain/kotlin/borg/trikeshed/sctp/SctpElement.kt
@@ -10,6 +10,29 @@ import kotlinx.coroutines.channels.Channel
 
 enum class SctpChunkType { DATA, INIT, INIT_ACK, SACK, HEARTBEAT, COOKIE_ECHO, COOKIE_ACK }
 
+/** Multi-homing path state for SCTP failover (RFC 4960 §6.4). */
+enum class PathState {
+    /** Path is actively used for data transmission. */
+    ACTIVE,
+    /** Path has failed heartbeats and is not used. */
+    INACTIVE,
+    /** Path has not yet been probed. */
+    UNKNOWN,
+}
+
+/**
+ * Per-path status for multi-homing failover tracking.
+ *
+ * @param address The destination address (host:port or IP).
+ * @param state Current path state.
+ * @param failures Consecutive heartbeat failures on this path.
+ */
+data class PathStatus(
+    val address: String,
+    val state: PathState = PathState.UNKNOWN,
+    val failures: Int = 0,
+)
+
 enum class SctpState {
     CLOSED,
     COOKIE_WAIT,
@@ -316,11 +339,50 @@ suspend fun openSctpElement(): SctpElement =
 class SctpElement(
    val streams: MutableMap<Int, StreamHandle> = mutableMapOf(),
    val associations: MutableMap<Long, SctpState> = mutableMapOf(),
+   val paths: List<String> = emptyList(),          // multi-homing: active path addresses
+   val congestionControl: String = "cubic"          // cubic | hstcp | rack — deterministic only
 ) : AsyncContextElement(), StreamTransport {
     companion object Key : AsyncContextKey<SctpElement>()
 
     override val key: AsyncContextKey<SctpElement>
         get() = Key
+
+    /** Per-path failover tracking — lazy-initialized from [paths] on first access. */
+    val _pathStatuses: MutableMap<String, PathStatus> by lazy {
+        paths.associateTo(mutableMapOf()) { it to PathStatus(address = it) }
+    }
+
+    /** Current primary path — first ACTIVE path, or null if none are active. */
+    val primaryPath: PathStatus?
+        get() = _pathStatuses.values.firstOrNull { it.state == PathState.ACTIVE }
+            ?: _pathStatuses.values.firstOrNull { it.state == PathState.UNKNOWN }
+
+    /**
+     * Mark [failedPath] as INACTIVE and return the next available ACTIVE path
+     * for failover. Returns null if no paths remain.
+     */
+    fun failover(failedPath: String): PathStatus? {
+        val current = _pathStatuses[failedPath] ?: return primaryPath
+        _pathStatuses[failedPath] = current.copy(
+            state = PathState.INACTIVE,
+            failures = current.failures + 1,
+        )
+        return primaryPath
+    }
+
+    /** Mark [path] as ACTIVE (recovery after successful heartbeat probe). */
+    fun recoverPath(path: String): PathStatus {
+        val current = _pathStatuses[path] ?: PathStatus(address = path)
+        val recovered = current.copy(
+            state = PathState.ACTIVE,
+            failures = 0,
+        )
+        _pathStatuses[path] = recovered
+        return recovered
+    }
+
+    /** All path statuses, indexed by address. */
+    val pathStatuses: Map<String, PathStatus> get() = _pathStatuses
 
     override suspend fun openStream(): StreamHandle {
         requireState(ElementState.OPEN)
